Split importers/base.py into a base/ package (thin ImporterBase + rich DatasetImporter)
Convert the single 1203-line base.py module into a base/ package:

- core.py: ImporterBase, the thin base every importer shares (metadata,
  dataset-name resolution, origin building, CLI wrapping, chunked-loading
  scaffolding, precomputed dicts, origin reload/display/resolve). Its run()
  is abstract and raises.
- dataset_importer.py: DatasetImporter(ImporterBase), which layers the
  source-spec / converter / ingestion pipeline and the list_records /
  fetch_record per-record hooks on top. Keeps the exact public API and
  behavior it had, so every spec-aware / service importer and all tests
  subclass it unchanged.
- specs.py: SourceSpec + spec parsing / converter-ingestion helpers.
- origin.py: origin-serialisation policy helpers + dataset-name field.

Migrate the 6 truly-thin importers (synthetic, pickle, local_files,
combine_datasets, demo, local_folder) onto ImporterBase so they no longer
inherit the multi-media machinery they never used. DatasetImporter remains
the documented public base, fully backwards-compatible for out-of-repo
extension importers.
